requirements.txt
langchain
newspaper3k
streamlit
tiktoken
